+++ CHANGELOG.md
@@ -1,9 +1,27 @@
+<<<<<<< HEAD
+=======
+## v0.2.0 (2026-07-29)
+
+### Feat
+
+- **training**: improve data processing and model training
+- **report**: add new functions and metrics
+- **download**: improve dataset loading
+
+>>>>>>> main
 ## v0.1.0 (2026-07-25)
 
 ### Feat
 
 - prepare repository for initial cz bump
 
+<<<<<<< HEAD
+=======
+### Fix
+
+- resolve remaining bugs in reporting and training
+
+>>>>>>> main
 ## v0.0.1 (2026-07-25)
 
 ### Fix